requirements.txt
aiogram>=3.15,<4
aiosqlite>=0.20
APScheduler>=3.10,<4
dateparser>=1.2
python-dotenv>=1.0
pytz>=2024.1
